E-commerce/dailyfresh/requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==1.8.2
django-redis==4.9.0
django-tinymce==2.6.0
fdfs-client-py==1.2.6
idna==2.7
itsdangerous==0.24
kombu==4.2.1
mutagen==1.41.1
Pillow==5.2.0
pycparser==2.18
PyMySQL==0.9.2
pytz==2018.5
redis==2.10.6
requests==2.19.1
six==1.11.0
urllib3==1.23
vine==1.1.4
